pyproject.toml
[project]
name = "guardrails-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "Vaibhav Jain",email = "vaibhav@aiplanet.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "guardrails-ai[api] (>=0.6.5,<0.7.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[tool.poetry.scripts]
main = "app.main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
